Move kube_legacy manager stub out of cmd/
